preFinal

+++ app/src/main/java/com/example/weather/Sys.java
@@ -5,7 +5,7 @@ public class Sys {
     public Integer type;
     public Integer id;
     public String country;
-    public Integer sunrise;
-    public Integer sunset;
+    public long sunrise;
+    public long sunset;
 
 }
